requirements.txt
google-cloud-logging
Flask
functions-framework